requirements.txt
Django==1.11.5
django-mptt==0.8.7
django-tastypie==0.14.0
psycopg2==2.7.3.1
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.2
six==1.11.0
